[Background Model] Use own QCache for thumbnails
Previously it would use an on-disk KImageCache which would never be invalidated.
Moreover, it just uses KIO::PreviewJob which does its own more clever (compares mtime and other properties) on-disk caching already.

It is using its own QCache rather than QPixmapCache to not jeopardize other caching for the potentially large wallpaper thumbnails
that are only relevant while the dialog is open.

BUG: 395447
FIXED-IN: 5.15.0

Differential Revision: https://phabricator.kde.org/D16095