[Device Notifier] Add a button to unmount all devices
Summary:
When at least two removable devices are mounted, a button shows
up that will allow to unmount all mounted removable devices.
This is convenient for removable drives with several partitions,
each of which have to be unmounted to be able to safely plug the
device out.

{F6330009}

FEATURE: 395644

Test Plan:
Plug and mount two devices.
Click on the new 'unmount all' button.

Reviewers: #plasma, #vdg, ngraham, broulik

Reviewed By: #plasma, #vdg, ngraham, broulik

Subscribers: abetts, broulik, ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D16212