[ksmserver] Rewrite Startup
Summary:
The code is effectively 3 startup phases each of which has some parallel
methods that need to be complete before moving onto the next phase.

The old code had a tonne of methods all of which tracking what state
we're in and trying to start the next method in the chain handling
things out of order.

This simplifies everything into 3 composite kjobs. A lot more classes,
but each one tiny with the flow more readable.

Code should be effectively the same. The startup sound is moved ever so
slightly earlier to be when phase 2 starts rather than just after
kcminit2 finishes (midway through the old phase 2), but no significant
behavioural changes here.

Test Plan: Logged in with session restore and without

Reviewers: #plasma, apol

Reviewed By: apol

Subscribers: apol, broulik, anthonyfieroni, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D16231