[ksmserver] Simplify call to kcminit
Summary:
kcminit had a slot to start processing something and a signal to emit
when it's finished.

For DBus that's quite bad practice as it can leave you hanging in the
case of a kcminit crashing. It's better to use the return of the
original call to signal when something is done.

From ksmserver we can then just use DBus timeouts set on the interface
than a secondary timer.

We also remove a pointless blocking call to kcminit introspecting
everything, if kcminit is not running it will return with an error
immediately which is no slower than checking first if it's not running
and faster if it is.

Test Plan:
Logged in
Session logs showed kcminit ran

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: zzag, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D16587