Add accessibility and set focus on the actual items in the window switcher
Summary:
For accessibility reasons we want to make sure that the QQuickWindow
thinks the focus is on the correct item. This will allow screen readers
to read which window is current in the window switcher.

Test Plan:
Running with accessibility inspection tools I see that the
right item has the focus with this change.

Reviewers: davidedmundson, chempfling, kwin

Reviewed By: davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Maniphest Tasks: T7034

Differential Revision: https://phabricator.kde.org/D16638