[logout-greeter] Use correct enum for logout hint to QML
Summary:
Libkworkspace has two duplicate enum values for the same thing, with one
marked as "//FIXME remove". My patchset is trying to kill that enum, but
we still use it for backwards compatibility in the L&F QML in the greeter.

In my split, I used the wrong one.

BUG: 401046

Reviewers: mart

Reviewed By: mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D16891