[Device Notifier] Keep the device label & icon after unmounting
Summary:
When unmounting a device, the device notifier keeps the information
of the device and its icon for a few seconds before it disappears.
But the binding on the label and the icon was making them go blank
before, leaving an emblem without an icon and a 'This device can
now be safely removed' notice without the name of the device
(although it did not always happen, this I am not sure why).

The problem appeared with commit 61b2b173e8d68623a3681fd27c0ad0348bf41191

Test Plan:
Unmount a removable device with the device notifier: for a few
seconds, the device name and icon should still be displayed

An example of the problem:
{F6442095}

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, filipf, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D17127