[Device Notifier] Fix transition animation
Summary:
In the specific case of having the device notifier opened
and inserting a device that has been configured to be
automatically mounted, the transition animation could be
interrupted resulting in the opacity being wrong.

Test Plan:
In the setup discribed in the summary, you should never
see something like the attached screenshot where the list
of action opacity is not 1.

{F6460038}

See http://doc.qt.io/qt-5/qml-qtquick-viewtransition.html#handling-interrupted-animations
for more information.

Reviewers: #plasma, #vdg, ngraham, davidedmundson

Reviewed By: #plasma, #vdg, ngraham, davidedmundson

Subscribers: broulik, davidedmundson, ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D17394