[kuiserver] Avoid double warning on terminated jobs
Summary:
When a job is terminated, it isn't necessarily removed from the job
owners queue to handle pending requests.

If a service unregisters and the job is already finished, we don't want
to mark it as an error.

Test Plan:
kioclient5 (when running with fixed interactive-ness) quits when there's an error.
We used to always show the client died error.

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D17711