[sddm-theme] Add buttons to username prompt to make it a full-fledged login screen alternative
Summary:
This patch adds the buttons Suspend, Restart, and Shut Down to the username prompt screen that appears once the Type User button is clicked so the user is able to suspend/restart/shut down from that screen without having to return to the initial login screen.
It also replaces the "Back" button with "List Users" (and an appropriate icon) on account of the latter being a more accurate description of what the screen goes back to once the button is clicked (and in case of future improvements, enables it to become a standalone login screen as well).

Test Plan:
Before: {F6605023}
After: {F6605026}

Reviewers: #vdg, #plasma, filipf, davidedmundson, ngraham

Reviewed By: #vdg, #plasma, filipf, davidedmundson, ngraham

Subscribers: abetts, davidedmundson, ngraham, filipf, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D18893