[sddm-theme/lock screen] Tighten clock and username shadows
Summary:
This patch tightens the shadows around the clock in the Breeze SDDM theme and lock screen so as to prevent the appearance of a diffuse dark glare that appears around the clock on top of very bright backgrounds.

In addition, it tightens the username/prompt shadows to match the shadows used in desktopcontainment.

Test Plan:
Username shadows, before:
{F6640921}
and after:
{F6640923}
(the new shadows take after the ones in desktopcontainment)

Before (white background):
{F6635960}
After (white background):
{F6641670}

Before (Next):
{F6635964}
After (Next):
{F6641671}

Reviewers: #vdg, ngraham, filipf, mart

Reviewed By: #vdg, ngraham, filipf, mart

Subscribers: mart, plasma-devel

Tags: #plasma

Maniphest Tasks: T10325

Differential Revision: https://phabricator.kde.org/D19325