[Login, Lock, and Logout Screen] Make the avatar background circle more subtle
Summary: Added in D16879, the background circle was designed to work on a blurless login screen. Since we have abandoned that idea, it's good to adjust it to our present needs. This patch lowers the opacity of the circle and removes the double border it created.

Test Plan:
Before:
{F6649241}

{F6649240}

After:

{F6649244}

{F6649243}

Reviewers: ngraham, #vdg, #plasma, rooty

Reviewed By: ngraham, #vdg, rooty

Subscribers: abetts, plasma-devel

Tags: #plasma

Maniphest Tasks: T10325

Differential Revision: https://phabricator.kde.org/D19478