[sddm-theme/lock screen] Fix "unable to assign [undefined]" error
Summary:
This patch fixes the "Unable to assign [undefined] to QColor" and "Unable to assign [undefined] to int" errors that occur in UserDelegate.qml (188:9, 187:9) and Clock.qml (53:13, 52:13, 60:13, 59:13), visible in sddm-greeter's output.

It does so by resetting the label style to Normal (instead of undefined). Once Text.Normal is set, the outline color is set as transparent (but that doesn't matter since there's no outline to begin with).

Test Plan: Works the same, looks the same, no error in output.

Reviewers: #vdg, #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D19577