[sddm-theme] Replace login button label with icon
Summary: This patch replaces the login button label "Log In" with an icon, places it to the right of the password field, and centers the combination of the password field and the login button. It also adjusts for the button's right border, which has a tendency to extend to the right and beyond the username input field by one pixel.

Test Plan:
Before:(user list screen):
{F6654659}
After (uer list screen):
{F6654661}
Before (user prompt screen):
{F6654662}
After (user prompt screen):
{F6654663}

Reviewers: #vdg, #plasma, ngraham, filipf

Reviewed By: #vdg, ngraham, filipf

Subscribers: davidedmundson, abetts, ngraham, filipf, plasma-devel

Tags: #plasma

Maniphest Tasks: T10325

Differential Revision: https://phabricator.kde.org/D19214