Image Wallpaper Slideshow - display the list of images that will be shown
Summary:
This shows all the pictures inside the folders added to the Folders list. I also tried to make single pictures excludable via a checkbox on the thumbnail. This is the first time for me programming with QT/QML/Singals-Slots and I tried to use as much existing code as possible. The thumbnail view is the same as for single images and I simply subclassed the listmodel. However even if I tried to do everything like the code for slidePaths it doesn't work correctly. The checking/unchecking of images only applies on restart of plasmashell. Maybe it's a single mistake that is easily spotted by a more experienced programmer otherwise if the thumbnail view is accepted I can also revert all the checkbox stuff.

FEATURE: 403703
FIXED-IN: 5.16.0

{F6595564}

Reviewers: #vdg, ngraham, davidedmundson

Reviewed By: #vdg, ngraham, davidedmundson

Subscribers: filipf, mart, alexde, davidedmundson, ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D18809