[notifications] Lift up notification content if one line of body text droops
Summary:
This patch lifts up the text content of notifications in which the combination of a single line of body text and its heading are taller than their icon and thereby prevents the notification content from drooping too far down and having too much of a bottom margin if it should become much larger than the icon. Seeing as the icon and its margins are never going to be taller than the content in question and the icon is always going to have 1 units.smallSpacing above it, this patch adjusts the height accordingly (0.5 units of smallSpacing above + 0.5 for the spacing below the heading + 0.5 below the content, accommodating the text) to achieve more even padding.
This patch should also improve padding with fonts greater than 10 pt in size.

BUG: 404407
FIXED-IN: 5.15.3

Test Plan:
Before (left) and after (right):
{F6616780}

Reviewers: #plasma, #vdg, broulik, ngraham, filipf

Reviewed By: #vdg, ngraham, filipf

Subscribers: cfeck, anthonyfieroni, filipf, ngraham, abetts, broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D19036