[Logout Screen] Show info that rebooting will enter firmware setup
When the flag to reboot into BIOS/UEFI setup is set, show a hint on the logout screen.
For simplicity and to avoid the layout from shifting, and since the user had to explicitly enable the flag
and probably know what they're doing, it is shown regardless of what action is highlighted.

Differential Revision: https://phabricator.kde.org/D19562