[service runner] throw out kde3 "compat" code
Summary:
Desktop files such as kde/foo.desktop are identified as kde-foo.desktop
because of the menu spec

```
If the directory contains sub-directories then these sub-directories should be (recursively) scanned as well. The name of the subdirectory should be added as prefix to the desktop-file id together with a dash character ("-")
```

the runner therefore broadly assumed that kde-foo.desktop is always
the desktop file kde/foo.desktop and since kde/ was only used up to KDE3
this assumption formed the basis for sepcial KDE3 compat handling. that is
of course wrong because a service might just as well actually have a
prefix and be actually called kde-foo.desktop rather than
kde/foo.desktop. Specific example: kde-gtk-config.desktop

seeing as this is ancient compatibility code that is actually wrong,
the easiest solution seems to be dropping the code entirely (and thus no
longer assume kde- has some special meaning).

(also I didn't find a way to get the underlying filename of a kservice,
 so making the code not wrong seems non-trivial)

CHANGELOG: the GTK+ settings module now correctly appears in krunner and kickoff
BUG: 383287

Test Plan: gtk kcm correctly shows up in both kickoff and krunner searches

Reviewers: broulik

Reviewed By: broulik

Subscribers: GB_2, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D19788