[KRunner] Avoid writing history if addToHistory is idempotent
Summary:
It is not uncommon to run the same command repeatedly. In this case,
the history is actually unchanged - the item is removed from the first
position, and prepended again.

Test Plan:
run the same command twice, config file is not rewritten
run a new command, config is updated

Reviewers: #plasma, broulik, apol

Reviewed By: #plasma, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D20383