Add SDDM theme login load feedback
Summary:
Currently, when you log in you can see a short freeze which looks bad. This patch adds some feedback by disabling and reducing the opacity of the controls after you begin the login (a busy cursor wouldn't work well, because it only changes when you move the mouse).
{F6800357}

Test Plan: Log in to a user.

Reviewers: #plasma, #vdg, ngraham

Reviewed By: #vdg, ngraham

Subscribers: ngraham, #vdg, plasma-devel, #plasma

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D20902