[PanelView] Update mask as last on resize event (& move one for consistency)
Summary:
The mask of a panel depends on things like borders set or window size,
but nothing else on it. So it should be updated as last in the queue of
things to update.
There might be still state changes not properly propagated, but this is
a move in the more expected direction.

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: apol, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21012