[Notifications] Avoid stale screenRect
AppletInterface does not emit change signals for screen properties, only ContainmentInterface does it.
Since a fix will only be in a future Frameworks version 5.16 won't depend on, mitigate the effect by not caching the property.
The property was only a lazy Connections replacement anyway.
Also, compress screen change updates to avoid flickery animation of the popup as you resize the panel.

CCBUG: 407504
BUG: 399257

Differential Revision: https://phabricator.kde.org/D21365