[System Tray] Ignore services that start with a colon (:)
Bus names that start with a colon (':') character are unique connection names rather than well-known bus names
you want to be listening for.
Since the same code path is used for initial setup as is for detecting services at runtime, we would needlessly
iterate the list of known dbus-activatable services as well as apply a regular expression on them for all unique
connection names wasting some cycles.

Differential Revision: https://phabricator.kde.org/D21451