[Notifications] Spawn new notification if one to replace doesn't exist
Since the Notification Server knows nothing about what notifications were sent, it tells us a notification was replaced and
uses the ID supplied by the application. Only in the model we then realize it didn't exist.
This means we will potentially reuse the notification ID sent by the application as we just increment a counter.
However, in practice I haven't encountered this issue before (maybe we just never noticed?) but I think this is an
acceptable workaround.

BUG: 407881
FIXED-IN: 5.16.0

Differential Revision: https://phabricator.kde.org/D21474