move SolidUiServer::showActionsDialog to hotplug dataengine
Summary:
it was the only remaining user of that dialog helper.
the connection between the two was fairly meh. the engine would invoke the
kded module over dbus, the module would then process the request, find that
only one desktop entry is specified, resulting in one action being found
and then running that action on behalf of the engine. this is overly
complicated for no additional gain. the engine is the sole user of this
functionality AND it already had knowledge of the service system, so
soliduiserver in this case doesn't actually add anything. it's just a more
involved call chain.

the useful code of soliduiserver is now in the hotplug dataengine directly
and the soliduiserver no longer supports the actions dialog. hotplug is now
the sole authority for hotplugging.

soliduiserver:
- include cleanup
- all classes no longer in use have been removed
- kdelibs4support is no longer a link target
- new link targets i18n + widgetsaddons (previously pulled in by
  kdelibs4support)

hotplug:
- hotplugjob now directly executes service actions using classes imported
  from soliduiserver
- no longer links qtdbus
- new translation domain plasma_engine_hotplug (for i18n call in imported
  service classes)
- new link against ki18n for that reason

Test Plan: running an action works same as before, internally it no longer relies on dbus/kded

Reviewers: broulik

Reviewed By: broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21761