drop kdelibs4support from filebrowser dataengine
Summary:
this engine is not used anywhere!

the engine used kfilemetadata to obtain metadata of a file. KFM however
is entirely defunct and does absolutely nothing without nepomuk.
that is to say: it is already broken for years and in fact not even
correctly sets the item.type key because of how the original code worked.

this change drops the kfilemetdata use altogether and instead sets the
item.type (which we can easily determine) to file when working with a file

in theory (I think) Baloo/File is the modern replacement for kfilemetadata,
but its properties are no longer "stringy" so we'd have to map a huge
enum manually into stringy keys for our representation. which at this time
seems not worth it, considering we (KDE) do not seem to use the engine
anymore and third parties would have had an entirely broken engine WRT file
sources as well, so I am guessing no one outside KDE uses it either.
all things considered I think simply rolling the engine back to the
minimum functional use is the best course of action here.

this effectively only changes the engine behavior to once again set
item.type=file on actual files while making it no longer require
kdelibs4support.

Test Plan:
in engine explorer requesting a directory gives unaltered directory data,
when requesting a file it correctly sets the type to file

Reviewers: davidedmundson

Reviewed By: davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21765