do not use legacy app_icon helper on klipper
Summary:
kde4_add_app_icon is fairly deprecated. additionally the call itself
was madness because it actually pulls the icon out of oxygen in the install
location. that is assuming the icon theme is even installed at build time
(which is probably always not true in our binary factorization
environments). lastly the app_icon helper is only useful to pick up
icons for osx and windows, neither are target platforms for plasma.

remove the app_icon call entirely until a use for it arises, which should
be... never... unless we decide that we want to replace explorer.exe ;)

Reviewers: davidedmundson

Reviewed By: davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21766