[interactiveconsole] Fix initial load selection
Summary:
QActionGroup has no signal for when an action within the group changes
checked state. Only when an action within is triggered.

Without calling trigger InteractiveConsole::modeSelectionChanged is not
triggered and m_mode is not updated.

BUG: 408787

Test Plan:
qdbus to launch the console
Execute button went to the right place

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21852