print a warning when invoking the scripting API's sleep()
Summary:
the sleep is implemented with a nested event loop which may result in
segfaults when something in the new loop stack accesses objects from the
original loop stack that are presently in a bad state because they were
not meant to be accessed (and only can because of the nesting).

to aid in debugging let's leave a trace of the fact that sleep was used
when in fact, using sleep is super risky.

Test Plan: warning gets printed when sleeping

Reviewers: mart

Reviewed By: mart

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21895