[Notifications] Make popup placement more resilient
The popup width is fixed, so don't rely on the actual dialog window width which might be bollocks.
More importantly, place the popups immediately and before they are shown so they show up
correctly and don't fly all over the place.
It can happen that the popup contents aren't readily laid out when we enter positionPopups() after
object creation, so don't reserve any space for the popup until it

Differential Revision: https://phabricator.kde.org/D22045