[Notifications] Ignore excess spam
When multiple identical notifications are sent in quick succession, discard the request and let it fail.
Previously it would always replace the last notification if identical, now it discards the request and only does so if it happened
within 1 second of the last notification to avoid discarding genuine identical notifications like when deleting two plasmoids of
the same type.

BUG: 409157
FIXED-IN: 5.16.3

Differential Revision: https://phabricator.kde.org/D22088