[Image Wallpaper Slideshow] Allow setting of different sorting orders
Summary:
Allows setting of other sorting orders like alphabetical or last modified date.
To enable this a new ProxyModel is introduced which handles the sorting and
filtering (as indicated by the checkboxes). This is backed by the slideshowModel
whcih previously as only used for the configutation. The lists of slides and  unseen
slides are dropped as now the slides that are shown are taken from the model.

FEATURE: 186181
FIXED-IN: 5.17.0

Test Plan:
{F6938259}
{F6938260}
{F6938261}
{F6938263}

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, msdobrescu, ngraham, filipf, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D22121