[Notifications] Keep new delegate hidden until room has been cleared
When a new notification comes in, we first displace the existing ones to make room and then animate in the new one.
However, due to the PauseAnimation the item appears with its full opacity ontop of the existing notifications that move out of the way.
Add a PropertyAction to ensure the item has an opacity of zero initially.

Differential Revision: https://phabricator.kde.org/D22760