Remove the ShellManager
Summary:
We had an overly complex and in practice unexploited system to detect
which shell to use,
This patch changes it to default to the desktop, still allowing to
modify which to use from --shell-plugin argument.

Test Plan:
Starts up fine, --shell-plugin/-p still works, which is what
bhushan told me is important for plasma mobile.

Reviewers: #plasma, davidedmundson, bshah

Reviewed By: #plasma, davidedmundson, bshah

Subscribers: bshah, davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D22874